# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Reactor
  version: 1.26.10-beta8
  license:
    name: BUSL-1.1
    url: https://mariadb.com/bsl11/
  description: |
    Auth, storage, functions, sites, health, and platform routes.
    Data stays on PostgREST at /data/v1.
servers:
  - url: http://127.0.0.1:18000
    description: Local Compose
...
paths:
...
  /auth/v1/factors/passkey/verify:
    post:
      tags: [auth]
      summary: Finish a password login with a passkey
      responses:
        "200":
          description: Session
        "401":
          $ref: "#/components/responses/Error"
...
components:
...
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error: {type: string}
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  securitySchemes:
    operator:
      type: http
      scheme: bearer
      description: Operator token
    service:
      type: http
      scheme: bearer
      description: Project service key
    user:
      type: http
      scheme: bearer
      description: User access token
    project:
      type: http
      scheme: bearer
      description: Anon key, user access token, or service key